+++ apps/client/src/app/lazy-data/lazy-data-keys.ts
@@ -49,6 +49,7 @@ export const LazyDataKeys = [
   'gatheringSearchIndex',
   'gcNames',
   'gilShopNames',
+  'gubalBnpcsIndex',
   'hqFlags',
   'hunts',
   'hwdInspections',