+++ apps/client/src/environments/patch-notes.ts
@@ -1,8 +1,12 @@
 export const patchNotes = `### Bug Fixes
 
-* **list:** better npc grouping logic.
+* **island-recipes:** fixed some recipes having ingredients x0 listed in database.
+* **island-workshop:** better groove formula in cowries prediction.
+* **island-workshop:** fixed state import not always importing properly.
 
 
 ### Features
 
-* **island-workshop:** you can now set your second rest day (first one stays hardcoded to D1).`;
+* **island-animals:** added drops to the animals table.
+* **island-workshop:** new button to generate a list for the current planning.
+* **island-workshop:** new checkbox to exclude items that require pasture materials.`;